pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "metalclaw"
version = "0.1.0"
description = "Sandboxed GPU-accelerated AI agent runtime for Apple Silicon"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "rich>=13.0",
]

[project.scripts]
metalclaw = "metalclaw.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/metalclaw"]
